[KGEN] Teach the elaborator about `kgen.inlined_call` (#4163)
This PR teaches the elaborator about `kgen.inlined_call`. An
`inlined_call` to a region is treated the same as `call_param` of a
region since the region callee is going to get inlined in both cases.
However, the operations in the region cannot be recursively pushed onto
the worklist because they use the parameter scope of the caller:
the ops in a non-isolated from above regions are processed before their
parent `inlined_call` by moving the calls to after its child ops in the
worklist.

When `inlined_call`'ing a function/generator/interface, an instance of
the interface is generated with the region inlined, but these instances
are not well-formed on their own. They contained code becomes
well-formed when the function gets inlined at the callsite. The
elaborator removes these transient functions at the end of elaboration.
A caveat is that an `inlined_call` to an interface ignores any
benchmark, because it is not possible to benchmark instantiations of an
interface that use regions not isolated from above.

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 701d300aba7e8c8e779df4a4c8b70f5cb598963d