[Python][Utils] Make `export-pr` use tested utilities (#4232)
This patch introduces the `get_uncommitted_changes` utility, and
modifies `utils/export-pr.py` to use it and a few other already existing
utilities for working with shell commands. This changes the semantics of
the script slightly: previously the script would fail only if a (tracked
or staged) file was modified, and now it will fail if the file was
modified, renamed/moved, or if a new file was staged but not committed,
etc. As before, untracked changes won't cause an error, they will be
just logged.

Thise patch also refactors the git utility tests as needed to reduce
test code duplication.

MODULAR_ORIG_COMMIT_REV_ID: 1edcb50e8967e0bcfc5ca7c5dd602525d66d61da