[All] Add initial cmake structure for defining modular versions (#4273)
We've increasingly grown a need to have some type of versioning for
the code base, so that we can detect when things change. This PR adds
in a very basic structure incorporating Major/Minor/Patch/Revision, where
the first 3 are currently stubbed to 0 (as we haven't defined any kind of
release schedule), and the latter is defined as the current git revision.

We will want to refine this structure over time, and for customer releases
generally won't want to include the revision number, but this should give
us enough to start working with.

MODULAR_ORIG_COMMIT_REV_ID: 0803040045cccd908cf18e6015027a876625451b