[Lit] Diagnose extraneous tokens at end of declaration. (#4295)
Remember the end of declaration cursor when completing the initial
pre-pass over a decl, and compare it at the end when the decl is
complete.  If there are extraneous tokens, then there is something
wrong with the declaration, which is now diagnosed.

This fixes Issue #4181

MODULAR_ORIG_COMMIT_REV_ID: 36d58d368c3d588c773c29822de0722c83362eaf