[Lit] Add initial support for type checking special functions. (#4312)
Python/Lit have a number of special functions like __init__ that
are required to be implemented in a specific way.  Python
implements these with a variety of internal implicit contracts
that are detected at runtime.

Lit implements and enforces these statically, and this patch adds
the mechanism to reason about these and an initial set of type
checking rules around __init__ as an example of this.  There is
more to do here though.

MODULAR_ORIG_COMMIT_REV_ID: 6b995773bdb7f7e900ed89e905fd8def93edb895