[Lit] Add support for `*expr` expressions, make 'self' be a pointer. (#4319)
This add support for `*expr` as an expression in Lit.  This is a
compatible generalization of the Python grammar, which uses `*expr`
in limited places to unfold a sequence.

We could eventually support this as a dereference (etc) operator on
non-sequence values like pointers, but I need this in the shorter
term in order to spell pointer types.  This is required because the
self member of a struct is always implicitly a pointer (at least
until we have a immutability model which seems like a long way away).

In practice, lit users shouldn't care about this, they should always
just write self without a type annotation like "def method(self, ...)`
and go on their merry way, but it is important for generality that
we can also spell this.

MODULAR_ORIG_COMMIT_REV_ID: 72039af500f545db8c03d5c5156a6e9a666b1688