[Lit] Increment towards having scopes for every user visible type. (#4329)
This expands our ability to model the magic behavior of a Scope
from being hard coded on name to being an enum set on construction
of the scope.

This adds a scope for None type, which is useful when emitting
None result types from defs (but isn't used yet).

While in the neighborhood, make the LITStructDeclOp::build helper
add a block for the body, since it is always required.

MODULAR_ORIG_COMMIT_REV_ID: b4ffe364dac9b3e17af06c90b39b113c3bbf1a06