[Lit] Big rename: Scope -> DeclAST, NFC. (#4331)
The "Scope" object does provide a namespace to do name lookup in, but
also contains a variety of other information pertinent to a declaration,
including its location and level of resolved-ness.  This makes it be the
AST representation for a declaration, which is discarded when IR
generation is done.  Rename it to reduce confusion about its purpose.

This generally renames the operations that back a declaration into names
like "op" and "attr" to reduce confusion.

MODULAR_ORIG_COMMIT_REV_ID: 7ec803fc9a97634a0e20e31027655f830ff6e612