[KGEN] Move `pop.variant.visit` lowering to `lower-scf-to-llvm`. (#4374)
The `pop.variant.visit` operation needs to be lowered at the same time
as SCF operations in order for whem to be nested in each other and still
lower correctly. This is due to DialectConversion recursing on parent
operations when block rewrite actions are performed AND that SCF ops
require single-block regions. `lower-scf-to-llvm` is where control-flow
lowering to CFGs should be done.

MODULAR_ORIG_COMMIT_REV_ID: 380b973cb9ff3623be5a2479323015729e8409e4