[KGEN] Remove single object caching and link at the LLVM level (#4294)
We currently lower every function to an object before trying
to produce a standalone object, but then immediately try to
raise back to an LLVM module. All of our compilation flows at
this point have the LLVM available, making the lowering to object
an unneeded cost (and it is a large one, especially for small functions
that don't really need that level of caching). This PR updates
standalone object production to link at the LLVM level when possible,
and removes the need for precompiled objects in many cases.

These changes led to a 3x speedup in non-cached compilation, and also
provides a nice reduction in the size of the generated cache (given we
no longer cache single objects, we still cache composite objects though).

MODULAR_ORIG_COMMIT_REV_ID: 753b21ec9c7125f9a6a3608c022b9aea6d58fe40