[M] Move #mop.buffer to #m.aligned_bytes (#4438)
This attribute is the 'inline' equivalent of the built-in dense_resource
attribute, and uses the same hex-string encoding for an uninterpreted
bunch-'o-bytes. However it allows the required alignment to be expressed
explicitly if required. I'd be happy to leave it in mop but we need to
respect the alignment in MLIRToBEF which requires it to move down the
stack.

The mop dialect uses this to support its notion of RCBufferRefs to
aligned underlying buffers.

We sure do seem to have a lot of ways of expressing array literals,
would be open to a better approach!

MODULAR_ORIG_COMMIT_REV_ID: b8ac40c6ca66ef319eba096ef14c7ef421edda0f