[KGEN] Add error test for incorrect usage of locally-defined parameter (#4445)
Parameter scopes are graphs, which means that is isn't possible to refer
to a parameter in a higher scope without giving it an alias first. Add a
test to make sure this results in an error.

MODULAR_ORIG_COMMIT_REV_ID: a75d6868d9bae8d18bcb4e88cf8c632ceb2afcba