[Lit] Change expression emitIR to compute and return ASTType's. (#4478)
This provides the mechanics for computing and propagating AST Types
for much of what we're doing.  The result type for call results
and fp/string literals are completely wrong in this patch, but this
is a step forward on plumbing.

MODULAR_ORIG_COMMIT_REV_ID: 790f176836287e95c235fd3f466c1f7d0f120df6