[Lit] Revamp + extend the emitted value hierarchy. NFC (#4527)
This makes it more composable and splits out a storage type that is
common across all the aggregate value operations.  This makes use of
ASTType being newly pointer size to add it as a member and makes it
possible to construct a MValue with a type, but nothing does this
yet.

This also adds 'type' as a magic declaration.

MODULAR_ORIG_COMMIT_REV_ID: 7962eff69fdcb8fcff6627604bc0a0c589e926ac