[KGEN] Fix tracking of inlined calls (#4544)
The elaborator was giving erroneous "non-inlined call to symbol
instantiated with non-isolated region" errors due to incorrect tracking
of which calls get inlined. Now the logic is that
`resolveCallInputParams` checks all the bound input parameters for
non-isolated regions. If any are found, it verifies that only inlined
calls (calls to regions or `inlined_call`s to symbols) are allowed to
bind non-isolated regions as parameters. It then sets a flag that is
forwarded to the elaborator.

Fixes #4519

MODULAR_ORIG_COMMIT_REV_ID: f6ac41d6c89d3152f52348d8df46be3dc9cb0513