[KGEN] Add a CallIndirectOp (#4628)
This operation is use to invoke SSA values of `!kgen.signature` type.
The LLVM lowering of this operation is TBD because we don't have a
runtime representation for signature-type values, which may capture
values.

Thsi operation is needed because Lightning has expressions like
`fooObj.barMethod(123, 456)` which is actually `(fooObj.barMethod)(123,
456)` in the sense that the `.` is an operator that returns a signature-type
result, and then `(123, 456)` is an operator that calls the
signature-type value with a list of arguments.

MODULAR_ORIG_COMMIT_REV_ID: 7fd3493381bd87a1ce20fdddf02cb71f23b87a39