[Lit] Change type emission to use emitIR value emission path. NFC. (#4658)
Now that value emission can handle full values of parameterized
type emission, we can use it as the backing path for
ExprEmitter::emitType.  This requires a bunch of other minor details
to be squared away:

1) None is emittable as both a type and a value, but we want
   "x = None" to resolve to a none value and `-> None` to resolve
   to a None type.
2) PointerType has a magic lowering to POP::PointerType but also has
   a struct declaration, so magic processing needs to adjust to
   support that.
3) The magic lowering of Pointer needed to be implemented.

This makes the emitType path in the expr nodes dead, which will be
removed in a subsequent patch (to avoid overloading this one).

MODULAR_ORIG_COMMIT_REV_ID: c72a9388ae26d853c376064253e8396510512d77