[Lit] Move FunctionType to be a struct with a result type parameter.  NFC (#4660)
This starts moving function type to be properly declared.  We don't
have variadics yet, but the most important thing to unblock is the
return type anyway.  Capture it as a type parameter, and propagate
that when reparameterizing a Function value with []s.  This also
renames it from signature to function since that is the language
level name for it.

This isn't enough to fix the /actual/ thing I want to fix - which is
the result of calls being typed correctly, but it is a step along
the way.

MODULAR_ORIG_COMMIT_REV_ID: 2c0a48d285424ed79b31fa633b394d6e54ac7218