[Lit] Revamp "self" l-value support. (#4679)
Self is a bit of a unique concept in our type system in that it is
passed "inout" for a mutation method like init.  We don't have a
mutation model quite yet, but we need to pass it by implicit pointer
so that member accesses to it can correctly address it.

Handle this by special casing pointer types as arguments (until we
get parameter conventions).  Also generalize our declaration model to
understand that declarations may be generalized RValues (e.g. like
attribute values) as well as LValues (like self on a mutating method).

This allows us to tie together a bunch of things to enable member
references in init and other methods to work right.

MODULAR_ORIG_COMMIT_REV_ID: e6373bcbf979b482fef693493023be87d55c990f