[KGEN] Add inline_asm Op to the POP (#4639)
This adds an inlined assembly construct to the POP dialect. This construct would enable target-specific assembly within kgen. This is done as a tiny wrapper over the existing llvm.inline_asm op, which in turn is a tiny wrapper around the inlined assembly expression functionaliy. No checks as to whether the assembly is valid or not, and developers should use this with caution and have their code predicated on the system arch.

MODULAR_ORIG_COMMIT_REV_ID: 6436e08a4c3932b4b28de715f65830448be9a5d1