[KGEN] Add align option to stack_allocations (#4641)
This adds an align option to the stack allocation instructions. This enables clients to be in control of the alignment during stack allocation. The alignment factor can be specified as an immediate value or via a meta parameter.

MODULAR_ORIG_COMMIT_REV_ID: fe5075f7a834d55954f065554945af64af8b17f6