[Support] Add boilerplate for HLCF Dialect (#4694)
This PR adds the boilerplate for the HLCF dialect. This dialect will be
placed in the Support directory despite being created specifically for
KGEN/Lightning because it is type system agnostic -- it can be used with
any generic MLIR code. This does not imply a desire to upstream the
dialect, however.

MODULAR_ORIG_COMMIT_REV_ID: 7410ca73d0cea56503c24282f414e18a25852872