[Lit] Make 'name' explicit for addDecl. (#4823)
This makes the caller pass in the name attr instead of digging it out
of the declaration pass in. This is important for mangled names, because
the symbol name will be mangled on methods, but we only want the base
name
in the symbol table.

MODULAR_ORIG_COMMIT_REV_ID: fa777d299dee6a1dd2490c413703c38325d85c40