[Lit] Add support for initialization syntax -> __new__ (#4825)
This teaches expression emission to know that "calling" a type is
actually an invocation of the static __new__ method on that type.
This allows us to initialize values of struct types, which is a
pretty core operation for working with types.

MODULAR_ORIG_COMMIT_REV_ID: d9daad0ace8fbc5e7a80a8a303ad29a85778fcae