[Lit] Ban __init__ in structs, we will use __new__ instead. (#4826)
__init__ will eventually be used for classes, and I expect it to be
a common error to reach for this in structs, so ban it with a nice
error message.

The rationale here is that the signature for init is a fundamentally
mutating operation, but we want to produce new instances as rvalues.
__new__ is a good fit for this, and __init__ isn't.

MODULAR_ORIG_COMMIT_REV_ID: 9f604f54bbdb18e00cc8699f2594282365690b27