[Lit] Introduce ASTType comparison reduce use of FullType, NFC. (#4840)
Now that ASTType is entrenched and we have a reliable conversion from
ASTType to IR type, we can stop carrying both around through the system
redundantly.  This simplifies the code and makes it easier to work with
the types because we don't have to keep both in sync in parallel.

MODULAR_ORIG_COMMIT_REV_ID: 6effa95898a0615abfa7db0a83a3315effa1da02