[Lit] Depriviledge PointerType, making it built in, but not magic. (#4843)
With other changes, we are able to depriviledge the pointer type
struct almost entirely - the only reason we have it built in is so
`&Type` can be sugar for `Pointer<Type>` (which seems like a
pointless optimization now and should probably be removed).  If
we decide to keep it, it can be put in the stdlib like the object
type.

This is done by making the ASTType of an LValue always equal to the
RValue type, so we don't have to have a way to spell the implicitly
wrapper type.  This makes working with this stuff more convenient,
deletes a bunch of code, and make room for computed lvalues (those
with getter/setters).

MODULAR_ORIG_COMMIT_REV_ID: dcfea55efe205b58ce351d4c657602a477e667c2