[Lit] Completely remove PointerType and & sugar. (#4871)
This was not actually needed for lvalues and is now just clutter.
Remove it, because it will eventually live in a standard library if
we need it.  The & sugar should also live in a library ideally.

MODULAR_ORIG_COMMIT_REV_ID: cd2365c5618e91bbc5ec5156b20e955baed49cab