[KGEN] Make lists fixed-size (#4912)
Variable-sized lists are essentially giving an implicit parameter on
every function that is the size of the list. This is extremely
problematic for the elaborator to deal with an honestly isn't something
that we want anyways.

MODULAR_ORIG_COMMIT_REV_ID: f4033391f6c5e3a2ddaa09acf6312bf1f741b914