[Support] Introduce `UNWRAP_ERROR_OR_SET` macro (#4955)
Instances of `ErrorOr<T>` sometimes need to be unwrapped and immediately moved into an existing variable (e.g. because it is passed as a function parameter). This patch introduces a macro to help with this, and refactors the implementation of `UNWRAP_ERROR` to reduce code duplication.

MODULAR_ORIG_COMMIT_REV_ID: 68119a5361bd244a3e39ed740495664eed47167e