[BEF] Move BEFContents down into Support/ML/CompiledFrameworkLabel (#4963)
[Fixes #4896]

In the switch to mop.model I needed a way to validate and decode the
'frameworkName' attribute and BEFContents is 1:1 with that notion. To
avoid the weirdness of a dialect depending on BEF directly, push that
notion down into Support/ML as the way we differentiate the 'framework'
flavor needed by the Modular runtime.

MODULAR_ORIG_COMMIT_REV_ID: b8d15276ed00acf490ee319ec55b281f6eb17657