[Support] Add inflate/deflate capabilities from/to `cache.symbol`. (#4978)
This patch adds support for inflating a symbol that's been cached and represented as a `cache.symbol`, as well as support for deflating a symbol and replacing it with a `cache.symbol`. This patch also adds tests for deflating, and roundtrip support.

MODULAR_ORIG_COMMIT_REV_ID: 0f1f7eb6b90082b205f3367dfbcef27368846328