[Support] Remove `cache.symbol`. (#4997)
Because of #4995 I decided to try and see what happens without it, and the implementation is cleaner *and* the IR is a little nicer. Let's go with this and deal with the implications of having no body explicitly in KGEN. This also means that we can cache anything that has a region, it doesn't have to be a symbol. Because we aren't replacing the op, we also don't have to deal with the symbol table at all. This patch also cleans up comments on previous patches that are simpler to implement now that we aren't dealing with an extra op.

MODULAR_ORIG_COMMIT_REV_ID: b0fe94eb5e0d699133c5c14235a67e93765dbb9a