[Lit] Allow attributes on __mlir_op synthesized operations. (#5135)
This allows one to bind attributes to operations synthesized by
the __mlir_op thing, which is a pretty useful thing to have.  The
values have to be encodable as Attributes in the standard type
system which is a bit limiting at the moment, but this is still a
helpful step forward.

MODULAR_ORIG_COMMIT_REV_ID: 1ac9606023233a17b5efca65068ac87a8bbd89c2