[Support] Allow one to clear the cache level between benchmark iterat… (#5127)
…ions

It is useful to clear the cache between benchmark iterations. The use
case ISC micro-kernels that can be affected by cache-hit vs cache-miss,
or for operations that operate on tiny memory. This allows one to do
that via an option to grt.benchmark as well as a flag in the
benchmarking C++ API.

MODULAR_ORIG_COMMIT_REV_ID: eeee051f269f1db766a76628b8b549aef61a7810