[Lit] Enable use of arbitrary attrs in operation operand list. (#5161)
Operations can allow arbitrary attributes in their attribute list -
they are not limited to TypedAttrs.  This gives full access to them
in this special position.

This enables us to build instances of the index comparison predicate
attribute, which is needed for index.cmp.  With this we can implement
all the comparisons on Int!

MODULAR_ORIG_COMMIT_REV_ID: 41163498b3a94236de6b67c688697952d299cbd5