[Lit] Enhance ExprNode with a getRange() method, NFC. (#5171)
This follows the clang/swift/etc convention of return the end point
of the range at the start of the end token of the range for simplicity
and efficiency.  These ranges are not used yet.

While here, rename TernaryOpNode -> IfElseOpNode for consistency with
its enumerator and python nomenclature, and rearrange it to follow the
source order.  Also add missing location information in calls and
subscripts.

MODULAR_ORIG_COMMIT_REV_ID: 6b379782318873dd5df8122930001ce0b5ff4bc2