[Lit] Introduce !kgen.litdeclref type. (#5182)
This introduces a new type that is similar to DeclRefType but is
used pervasively by the Lit dialect and is lowered to DeclRefType
when semantic analysis and other lit lowering happens.

This type is different than DeclRefType in two ways:
1) It allows partially bound parameter types (like ASTType) which
   haven't yet had full inference applied to them.
2) It allows nested symbol references for declarations.  Lit doesn't
   currently support nested structs, but should, and the flattening
   transformation that lowers LIT decls to flat decls with mangling
   should also lower these out.

It is unfortunate that this has to be added to the kgen dialect,
I would much rather it be a !lit.declref type, but not doing this
breaks that POP dialect pretty printing, which is pretty important
in practice.

MODULAR_ORIG_COMMIT_REV_ID: b606c539f93c114810815dbe92864b61077beee6