[Lit] Track symbols for decls, and force ASTType clients to pass shared state. (#5186)
This changes DeclResolver/SharedState to track a map that inverts
the symbol -> ASTDecl mapping, allowing efficient lookup of decls
by their MLIR symbol.

As a related but different step, this changes all (but one) client
of ASTType::getDecl to pass in shared state, allowing a future
patch to use this map.  Some cleanup will be required to fix the
last case.

The testcase change is due to a minor logic bug fixed by this that
was using || instead of &&.

MODULAR_ORIG_COMMIT_REV_ID: c39ac5397bb7fda81975abb7236ee911b687d4dd