[Lit] Gut and replace the implementation of ASTType (#5191)
With the introduction of LitDeclRef type, the reliability of symbols,
the
symbol -> decl map, and other related refactorings we can gut ASTType.
Instead of needing a parallel representation of types, we can use the
MLIR
type system again.

This patch is a step towards unifying types into the metavalue system
and
consistently representing them as attributes.  This will allow a lot of
downstream simplifications in follow-on patches and eliminates the
parallel
uniquing structures that would otherwise be needed.

MODULAR_ORIG_COMMIT_REV_ID: b3092899d72ceee26dcad20994eb19b10467d191