[DebugInfo] Add a new debug info dialect (#5107)
The Debug Info dialect contains attributes, operations, and
types for representing debug information in MLIR. The dialect
is intended to be extensible, both in supporting different types
of external debug information formats and different types of
frontends. This dialect is intended to compose directly with any
other, making as few assumptions about the abstraction level as
reasonably possible.

The core structure of the dialect is a pragmatic evolution/cleanup
of the constructs built in LLVM over time, in addition to support
for "unresolved" debug information (to enable progressive lowering
of debug info alongside the IR it represents).

This PR adds the essential attributes, operations, and types to
represent programs we have today. There is a lot of additional
debug functionality that we can build up overtime, but this is
kept trimmed for now. Followup PRs will handle integration
into our stack and lowering to LLVM.

MODULAR_ORIG_COMMIT_REV_ID: 7d6466292c6cd1f9a992a74bbf5e811869f609c8