[Lit] Give `lit.func` ops parameters corresponding to context. (#5344)
It is possible to bind to a method in another type, and those
bindings need to be able to specify the values of parameters that
exist at the type level.  Support this by making them explicit
at the IR level.

This has the added benefit of defining away a possible class of
shadowing error by making it rejected structurally.

MODULAR_ORIG_COMMIT_REV_ID: 3bc5f406dc1e19e605837ce175b1522a824cabfe