[KGEN] Change KGENCallOpInterface.getSignature() to elide input params. (#5383)
Asking for "the signature" of a call is a little bit weird - it is
computing the effective callee of a call.  This callee doesn't have
input parameters specified in a call because they are bound by the
call itself.  Handle this by just dropping them from the result of
getSignature(), simplifying clients and making the codebase more
consistent.

MODULAR_ORIG_COMMIT_REV_ID: 7d5a71c393699981cbbd698ea5756dee973db111