[Lit] Type check parameterized method references correctly. (#5384)
Now that SignatureType is behaving, this changes emitFuncReference
to perform substitution of parameters based on the base access type.

Achieving this requires two things:
1) Collecting the "full" signature of a function for its symbol,
   which includes any contextual parameters from an enclosing struct.
2) Specializing the signature as a whole based on any available
   parameters.

For example, given an operator like `[T](Scalar[T], Scalar[T])`, we now
correctly form a reference to the symbol as having the full type above
and when T is found to be F32, we specialize the whole thing to
`(Scalar[F32], Scalar[F32])` instead of leaving the second argument as
Scalar[T].  With this change, a bunch of stuff now works.

MODULAR_ORIG_COMMIT_REV_ID: 18a44902c319ad8bceb163262fd94c8b2dbe12f4