[Lit] Rename Scalar2 to SIMD and parameterize on element type. (#5390)
This expands out the Scalar2 type to be a true SIMD type, implements
more methods and comments out some that are not implemented.

This also cleans up the inplace operators to invoke the non-inplace
operators instead of the implementation methods (just as a tidying up).

The SIMD type is still implemented in terms of MLIR parameter types,
which is fine for parity with kgen but not good enough in the long term.
It also drives the need for type aliases!

MODULAR_ORIG_COMMIT_REV_ID: 3aa733277c8da21cac0e9c32c62bcb3abf02fa48