[Lit] Rewrite call generation logic. (#5395)
Instead of emitting each subexpression in a piecemeal way, forming
'a.method' into a partial apply (and typically completing it),
emit calls using a new CallableValue abstraction.

The problem with the previous approach is that it assumed that each
subexpression could be materialized as an rvalue.  This was true for
non-parametric cases, but not true for methods that bind input
parameters: these cannot be partially applied without the params.

It turns out that this change will also be useful for overloading
in the future, as an overload set isn't usefully materializable as
an rvalue either.

This change is also useful in that it won't make partial apply
operations except when the user explicitly partially applies in their
code.  This will lead to more predictable codegen and can easily
emit a direct lit.call operation now, which will make it easier to
read the generated IR.

MODULAR_ORIG_COMMIT_REV_ID: 90fd8e54e6fa3e544dd18be419a6bfaeb3b10256