[Lit] Refactor CallableValue to clean up internals, NFC. (#5396)
This tightens up the internal implementation of CallableValue to make
the 'direct' call more explicit and to only form LITSymbolConstantAttr
when all of the bindings are formed.  Previously, we used
LITSymbolConstantAttr for the 'in progress' binding constructions, then
transformed it right before calling it.  This makes it more clear that
the maintained signature type is unremapped until the symbol is formed.

MODULAR_ORIG_COMMIT_REV_ID: 8713122e9d78ea9480b454508070c6512883f32e