[Lit] Expose kgen.param.assert to the standard library. (#5397)
This adds a new 'assertParam' function that exposes kgen.param.assert
to Lit developers (better name welcome).

Doing this requires a few infra enhancements: we need the ability
to specify a no-result __mlir_op, so we want to be able to emit
list expressions to parameters.  We're not ready to do this, and
unfortunately ArrayAttr isn't typed so we can't hack that, so just
handle the zero result case for now.

This also exposes a problem whereby we can't make the message be
parametric.  I will file an issue about that.

MODULAR_ORIG_COMMIT_REV_ID: de365991b19e91a4834d3ef3efcc76ce143da82f