[Cache] Add `cachedTransform`. (#5328)
This patch adds a generic way to provide a transform and have it cached. Users should be cautious about the granularity of that transform and ensure that the caching actually helps speed up their application.

We provide an overload of `cachedTransform` that takes an mlir::PassManager - we anticipate this will be a popular way to execute cached transforms, and it's very useful for testing.

MODULAR_ORIG_COMMIT_REV_ID: aac59f2a183e83e2dbdc038a9432a3db83dbb303