[KGEN] Generalize kgen.struct.decl to subsume lit.struct.decl. (#5479)
This extends kgen.struct.decl to allow it to hold other operations,
and to be a symbol table scope. This allows is to subsume
lit.struct.decl which allows us to de-duplicate a ton of operations
that works with these two forms of structs.

MODULAR_ORIG_COMMIT_REV_ID: 826cca7baa5cde09143c9a22463bfd1047cbb7b8