[Lit] Specialize 'ASTDecl::resolvedType' to 'selfType' for structs, NFC. (#5518)
The only load bearing use for this field is in structs which acts as a
cached type for the self type, and an override for the magic types.
Eliminate its use for other declaration kinds and make the accessor
name more specific to reflect this.  I'd love to eliminate this field
entirely.

MODULAR_ORIG_COMMIT_REV_ID: e7ba09aac5b6d95bbe24487fa943592d20665b77