[Cache] Set up `ConstantHashAttr`. (#5419)
This attribute is currently unused, but it will be used for caching large blobs inside regions. For that reason, regions are now parametrized on symbols *and* contained ConstantHashAttrs.

The RegionHashAttr is explicitly not recursive as a simplification. We will always need a root hash that cannot be parametrized, and it would be possible to make RegionHashAttr contain sub-regions, but the use case for this has yet to materialize so for simplicity we restrict the nesting to RegionHashAttr -> ConstantHashAttr.

MODULAR_ORIG_COMMIT_REV_ID: 8562b3cf9ecc7197067282cb2c105352cb3252c8