[Lit] Add ZAP dialect, diagnose recursive reference errors better. (#5562)
When diagnosing a recursive reference, emit a note saying where
the first reference was.  This makes it much more clear what is
going wrong in weird cases.  To make this work, enhance
declsCurrentlyProcessing to keep track of the locations of a reference,
and fix resolveAll to pass the location of a decl instead of start of
source file as the location.

MODULAR_ORIG_COMMIT_REV_ID: dff9ac501d48753db33b489975e2ed516456cc80