[Cache] Generalize the key passed into `cachedTransform`. (#5573)
This makes it easier for users to create more complex/in-depth keys for use in the cached transform. For simple cases like strings, it amounts to ~2 extra lines of code, but for complex cases it will be much simpler.

MODULAR_ORIG_COMMIT_REV_ID: 9da3589b6691f3eaa74ccf75b0870d32ce61b581