[MO][MOP] Implement basics of `mo` -> `mop` conversion (#5521)
This patch implements the basic infrastructure for converting the `mo` dialect to `mop`. In particular, this patch delivers the minimum scaffolding needed to lower an empty `mo.graph` to `mop.model`. The implementation is based largely on that of `mtfl-to-primitives`, with all the TFLite specific details removed.

Closes #5434.
MODULAR_ORIG_COMMIT_REV_ID: 749dab26f61a3bb3fcf7f4ad3553218d1c4afd3d