[Cache/KGEN] `cachedTransform` doesn't need to deflate/inflate ops. (#5611)
`cachedTransform` should not require a regionCache in the most generic
case - the decision of whether to inflate or deflate an op lies with the
implementer.

This also orthogonalizes the transform cache and the region cache
nicely, and means the elaborator only needs the transform cache.

MODULAR_ORIG_COMMIT_REV_ID: 1d1900ec852117c544445baeb8f39bd9938cef78