[Cache] Use AnyAsyncValueRef rather than LogicalResult for op inflation/deflation and cachedTransform. (#5675)
This changes the inflate/deflate APIs to use AnyAsyncValueRef - and therefore the AsyncValue error reporting - to clean up the APIs. This means that the result of cachedTransform is a Chain (if the transform was not run, so a cache hit) or the result of the transform itself.

MODULAR_ORIG_COMMIT_REV_ID: f50ba12e87d3ccaeddf2819dc08a289b9f2deed1