[KGEN] Used the new cachedTransform logic in the ObjectCompiler (#5665)
This is significantly cleaner, and also allows for us to asyncify the
internal compilation. As part of this transition though, I removed
the caching in lowerToLLVM because it isn't really used by anything
and it isn't clear that this is the right level of granularity to cache. We
can add it back if it ever becomes necessary.

MODULAR_ORIG_COMMIT_REV_ID: 448b358e45338fff1ed3ee25d7ebd1447d63c002