[Lit] Now that SignatureType has conventions, use it in Lit. (#5751)
This uses parameter conventions in lit to distinguish between
something that happens to have pop.pointer type and something that
is actually a by-ref argument.

This allows us to build abstractions around things like
index-to-pointer casts etc that are used widely by kernels.
This also is a more principled design that is easier to reason
about and extend in the future.

MODULAR_ORIG_COMMIT_REV_ID: 889e7e277b0647c568281505e92592d33181d74d