[KGEN] Remove the recursion checker from the Elaborator. (#5753)
This removes the recursion checker from the elaborator, and replaces
it with a depth test in the expander.  This catches infinite recursion
as well as the previous test, but allows cases where there is a base
case that cuts off the infinite recursion.  This is a lot more flexible,
particularly for things like "variadic templates".

This fixes Issue #5738

MODULAR_ORIG_COMMIT_REV_ID: 058915a381c60c8385e22806c074db85b5c1d3c5