[GenericML] NFC to bring BufferRef, Context and ContextState together (#5767)
We're on somewhat thin ice when it comes to all the buffer borrowing and
mutability we're doing when mixing TFLite and MGP tensors and their
respective contexts. I'd like to firm that up with a notion of scope
(not yet sure if it can be a debug-only concept, I suspect it is needed
in release when making copy vs borrow decisions in tflp.from_mtensor).

But before tackling that, this PR brings the MGP buffer, tensor and
context concepts together into GenericML/Support/ so I can talk about
this without breaking the layering. Also switch to 'GML' namespace so
can use throughout support.

Obviously in the future we'll want to support prims which have nothing
to do with tensors or genericml but nevertheless can make use of
buffers, context and other 'generic' MGP stuff. However at that point
we'll need to split both MGP-the-dialect as well as the runtime. So I
don't think putting BufferRef next to Tensor makes that future split any
harder.

Part of #5369.

MODULAR_ORIG_COMMIT_REV_ID: 340298251a9da672100be1a369ac48be2f23e734