[Support] Remove SIMDMatrix from repo (#5796)
SIMDMatrix is useful in principal, but the clang implementation is so
limited that it does not even make sense to use the extension.
Furthermore, it is complicated to use with a bunch of special cases
(e.g. only supporting column-wise order). This removes the SIMDMatrix
altogether. If we need the functionality, then we can expose the
relevant functions from the KGEN implementation which is a lot more
robust and powerful.

MODULAR_ORIG_COMMIT_REV_ID: 8f01cda4c4bac55e3ce6f361d9a35f80df6cb595