[Support] Add current duration to the state of the microbenchmark (#5833)
This adds the current duration to the state of the microbenchmark. This
allows the user to get the current duration of the benchmark via the
epilogue function. The motivation for this is to allow the user to
early terminate the benchmark if the current duration exceeds a certain
prior known value. E.g. when performing kernel search it makes no sense
to continue benchmarking if the current duration is more than 2x of the
fastest duration.

This is based on idea that @mvz suggested.

MODULAR_ORIG_COMMIT_REV_ID: 1916b68a2150b97c5030c47cec415952a8169daa