[KGEN] Terminate the benchmark if the duration exceeds the fastest time (#5834)
This terminates the benchmark at any point and does not wait for the
min-time to be reached within kernel search. This is useful because
we are only benchmarking the kernels to find the fastest one, and
a kernel is not going to get faster if you run it more times.

MODULAR_ORIG_COMMIT_REV_ID: 7a8e9b8daf800d7b7270749b327b9f2cea5d1bad