[KGEN] Split KGENDeclInterface into two interfaces (#5852)
This PR splits KGENDeclInterface into two interfaces: DeclInterface and
FuncInterface. DeclInterface denotes operations that declare a new
parameter scope. DeclInterfaces can declare parameters for the new
scope and constraints over those parameters, and that's it. StructDeclOp
is the only operation that is only a DeclInterface, which makes sense
because it does not define a callable thing with a signature.
FuncInterface, on the other hand, is an operation which defines
something with a signature. It is required to also be a DeclInterface.
This cleans up the mish-mash that was KGENDeclInterface. Further
cleanups can occur when the next LLVM integrate lands and *ArrayAttr are
implicitly convertible to ArrayRef.

This PR also "fixes" some weird edge cases with SignatureType by
allowing SignatureType to not be isolated from above, which allows the
signature types of functions to be verified through the same path as
other signature types.

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 34f27cc39cdac059c104ac0bfe8c8caefc566b0f