[Lit] Implement trivial implicit conversions in call arguments. (#5899)
When trying to pass an r-value to a function with a type that doesn't
match, attempt to invoke its new method on the expected type to convert
the operand to the expected type.  This eliminates a bunch of casts,
e.g. the need to say F32(1.0) everywhere.

This is limited and not great for a few reasons:
1) We don't support overloading, so lots of types aren't convertible
   from all the things they should be, e.g. you can't pass "1" to
   something that expects F32, because F32 can't be created from index.
2) This doesn't "check to see if we can invoke __new__" it force applies
   it on a mismatch, which leads to poor QoI.
3) This doesn't fix things that need radd.

Each of these will be addressed in subsequent patches though.

MODULAR_ORIG_COMMIT_REV_ID: c5470f62b49d7d019b098482e6421fada0772d83