[Lit] Improve QoI when missing a token at end of line. (#5906)
A common problem with parsers is that they report errors at the point
of the current token when they are missing one.  This makes sense when
the token is in the middle of a line, but not when the token is on the
start of the next line - the general problem is that someone forgot to
put the punctuation on the previous line.

Improve this by having the parser back up to the end of the previous
line in this case.

Before:

```
modular/KGEN/test/parse-lit/lit-crashers.lit:1 offset :7:3: error: expected ':' in function definition
  fn f():
  ^
```

After:

```
modular/KGEN/test/parse-lit/lit-crashers.lit:1 offset :4:33: error: expected ':' in function definition
  fn __new__(a: __mlir_type.f32)
                                ^
```

Not even clang knows this trick :).

MODULAR_ORIG_COMMIT_REV_ID: dbffbdd09e332270dcfdb205a48254ed47e454c7