[KGEN] Make FnEffects and ValueInputConvention MLIR enums (#5892)
This auto-generates the stringify and symbolize functions for these
enums. Also, this drops the logic where the first convention is assumed
as `none` if if is `byval` or `byref`. It's not important since I will
change the syntax for conventions in the next PR.

MODULAR_ORIG_COMMIT_REV_ID: a13d6e2ef1de5ee97d6d1f16cc02f107f3adab90